create measures-layout
created files:
measures layout.js and
utils/filter-functions.js
